+++ src/common/index.js
@@ -1,5 +1,7 @@
 const replaceIf = require("./replaceIf");
+const keyPredicate = require("./keyPredicate");
 
 module.exports = {
-	replaceIf
+	replaceIf,
+	keyPredicate
 }
