+++ src/index.js
@@ -1,9 +1,11 @@
 const date = require("./date");
 const graph = require("./graph");
 const math = require("./math");
+const io =. require("./io");
 
 module.exports = {
 	...date,
 	...graph,
-	...math
+	...math,
+	...io
 };
