+++ test/artifacts/textFile.txt
@@ -1,0 +1,1 @@
+This is some plain text in a text file.
