+++ test/io/file.js
@@ -1,0 +1,1 @@
+const file = require("../../src/io/file");
